Login is pretty much done. Access to UVA eservices balance is done.
Issue is considering people who have a slow internet connection, because it's asynchronous so I have a delay... Maybe it might better to make sure that a specific page is loaded before moving onto the next one? I need to work on keychain next because I keep having to cover my screen everytime I work on the code lol. Keyboard covering textfields is also an issue